Add files via upload
The circuit diagram for constructing an infrared thermometer using the MLX90614 sensor, Arduino Pro Mini, OLED Display, and a battery is detailed in attached image. This setup utilizes the Arduino Pro Mini board due to its compact and portable design. There are two versions of the Arduino Pro Mini: one operating at 5V, 16MHz, and the other at 3.3V, 8MHz. For this project, the 3.3V, 8MHz version is ideal since the device is powered by a 3.7V Lithium-Ion battery. The RAW pin of the Arduino Pro Mini is connected directly to the battery's VCC pin via a push switch.